+++ src/qwr/text_helpers.cpp
@@ -67,9 +67,9 @@ std::optional<UINT> DetectCharSet( std::string_view text )
     }
 
     constexpr int maxEncodings = 2;
-    int encodingCount = maxEncodings;
+    int encodingCount = static_cast<int> ( maxEncodings );
     std::array<DetectEncodingInfo, maxEncodings> encodings;
-    int iTextSize = text.size();
+    int iTextSize = static_cast<int> ( text.size() );
 
     hr = lang->DetectInputCodepage( MLDETECTCP_NONE, 0, const_cast<char*>( text.data() ), &iTextSize, encodings.data(), &encodingCount );
     if ( FAILED( hr ) || !encodingCount )
